Organize abstraction methods
